requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
bcrypt==4.1.2
PyJWT==2.8.0
python-dotenv
